Revert "Increased interaction queue size from 16 to 25 and cut the kMaxMinutesRemainingNotToAutosolve by a third"
Not fully sure, but thinking it was this change, but lately when something gets confused the interaction queue endlessly repeats showing items.  The 16 limit seems to be hard-coded in the game's code (could never have 25, it limited me to 16 even with that change) and the endless repeat and flashes might've been caused from the other change.

This reverts commit c07b0adafde584f32f201852665d1bfc4a811d92.